complete rewrite
Switched focus of script from turning ppp0 on and off. Instead, ppp0 is always on and the script focuses on switching default route context based on a series of tests.

Check default route count first:
if count > 0, proceed
if count = 0, add default route to primary WAN interface. Wait for next check to test...as there would be no network connectivity in this state anyway.

if ping succeeds on eth0
 -- and default route is to eth0, exit (all is well)
 -- and default route is to ppp0, delete it (fall back to primary)
 -- and there are multiple default routes, make primary WAN default route

if ping fails on eth0
 -- and default route is to eth0, move to ppp0 (failover)
 -- and default route is to ppp0, exit (all is well)

*CAVEAT* you must know your system routes and adjust as necessary. If you have any specialized routes happening, this script will break things.